apps/glusterfs: move db cmd ops to new file
